refactor: Re-organize domain.model.auth.
