+++ app/src/main/java/no/usn/mob3000/ui/screens/auth/CreateUserScreen.kt
@@ -26,7 +26,7 @@ import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.flow.Flow
 import no.usn.mob3000.R
 import no.usn.mob3000.Viewport
-import no.usn.mob3000.domain.model.auth.RegistrationError
+import no.usn.mob3000.domain.model.auth.error.RegistrationError
 import no.usn.mob3000.domain.viewmodel.auth.RegistrationState
 import no.usn.mob3000.ui.components.Loading
 import no.usn.mob3000.ui.components.auth.Error
